server/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.1
flatbuffers==24.3.25
frozendict==2.4.2
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.0
keras==3.2.1
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
multitasking==0.0.11
namex==0.0.8
numpy==1.26.4
openai==1.20.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
peewee==3.17.2
protobuf==4.25.3
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
yfinance==0.2.38